Ip6

+++ src/musl/openat.cpp
@@ -2,12 +2,12 @@
 #include <sys/stat.h>
 #include <fcntl.h>
 
-static int sys_openat(int /*dirfd*/, const char* /*path*/, int /*flags*/, mode_t)
+static long sys_openat(int /*dirfd*/, const char* /*path*/, int /*flags*/, mode_t)
 {
   return -ENOSYS;
 }
 
 extern "C"
-int syscall_SYS_openat(int dirfd, const char *pathname, int flags, mode_t mode) {
+long syscall_SYS_openat(int dirfd, const char *pathname, int flags, mode_t mode) {
   return strace(sys_openat, "openat", dirfd, pathname, flags, mode);
 }
